# Move structs to classes, ~1.10x faster Ripper.lex