@mame 2021-11-09
Thank you for your contribution! I like this change.

I confirmed the performance improvement with rdoc, a macro benchmark that uses Ripper.lex.

Without this patch:

```
$ rm -rf .ext/rdoc && time ./miniruby -I./lib -I. -I.ext/common  ./tool/runruby.rb --extout=.ext  -- --disable-gems "./libexec/rdoc" --root "." --encoding=UTF-8 --all --ri --op ".ext/rdoc" --page-dir "./doc" --no-force-update  "."
...
real    0m18.468s
user    0m18.060s
sys     0m0.400s
```

With this patch:

```
real    0m18.029s
user    0m17.691s
sys     0m0.336s
```

The speed up ratio is not so big (about 2%), but I ran them a few times to make sure it wasn't an error.

I understand this improvement as follows: (1) `tok.pos` is faster than `tok[0]`, and (2) attr_reader is faster than struct field access.

In regard to the incompatiblity,

> 2. Perform a deprecation cycle where we deprecate the Struct methods (such as `[]` and `each_with_index`).

I vote for this. IMO, it would be enough to mimic only `[]` for almost all cases. (I guess you mentioned `each_with_index` because of [this hunk](https://github.com/ruby/ruby/pull/5093/files#diff-00cab56d5cd91c959d4b06968b45aa010de48631b54a18593aabca9c261d075eR169), but I believe this code is a bug that just happens to be working well :-)