@schneems 2021-11-11
@mame thank you for the consideration. I updated to add `[]` method back in along with a deprecation https://github.com/ruby/ruby/pull/5093/commits/a66b299360168399f8684573d5646d1ed8740528. I've marked this as ready to review. Please take a look.