@k0kubun 2021-11-18
No objection to optimize `Ripper.lex`, which I'm a heavy user of, but you might also like https://github.com/ruby/ruby/pull/5131.