@mame 2021-11-18
@schneems Sorry for my late action. As k0kubun said, ko1 has improved the performance of field access of Struct. Could you measure the performance again?