@schneems 2021-11-19
## Master performance

I checked the performance of master after that commit versus 3.1.0-preview1 and it looks like master is slower by ~1.8% here are my raw numbers and benchmark scripts: https://gist.github.com/schneems/1edfff044a082ca61f962066541dfdb4.

I don't know if there is something else with `make runruby` that might be affecting the results, or perhaps some other change to master is slowing those results down.

## This patch

I am not sure how to reliably integration-test this patch against that C change. Running `make runruby` with different branches doesn't appear to re-compile the struct code. I could perhaps `make clean` between changing branches but that would add far too much time to produce 50-200 iterations.

I did however run the micro-benchmarks after pulling from master and re-compiling. Here are the results: https://gist.github.com/schneems/e1faf9bd43c626536f3631f8e6506583

The struct access is indeed faster now. It's being reported as "about same difference" as the class access. It actually looks slightly faster but it's within the margin of error.

Both class creation and `to_a` are still faster using a class instead of a struct:

```
Comparison:
     class creation :  3361272.1 i/s
     struct creation:  2770825.9 i/s - 1.21x  (± 0.00) slower
```

```
Comparison:
          class to_a:  9602229.1 i/s
         struct to_a:  6825047.7 i/s - 1.41x  (± 0.00) slower
```

I think there is a benefit to speeding up those two operations.