@schneems 2021-11-19
I was curious about how often each of these operations are called in `Ripper.lex`. I annotated the sourced code with a simple count of a global (https://gist.github.com/schneems/957703e08a183ddd25e632a0d0315753) this is what I got:

```
$ ruby -Iext/ripper/lib -rripper -e "Ripper.lex(File.read('array.rb')); h = { create: \$count_create, to_a: \$count_to_a, access: \$count_access}; puts h.inspect"
{:create=>424, :to_a=>212, :access=>212}
```

So it looks like create is called about 2x as much as `to_a` while `to_a` and access operations are  dead even.

Based on that call count info and the micro benchmark I posted previously I think this has roughly the same 1.10x faster impact on `Ripper.lex`, however it will not speed up access for places that use the internal Elem API (such as IRB and `dead_end`) because Koichi's patch made struct access comparable to regular object access.

It is my experience (with `dead_end`) that the bulk of time is spent generating the lex information rather than in consuming it (however you and @aycabta have more first-hand knowledge there with IRB).