@schneems 2021-11-19
I also annotated the tests in `test/irb/test_ruby_lex`

```
diff --git a/test/irb/test_ruby_lex.rb b/test/irb/test_ruby_lex.rb
index 32463f7fb6..dc500e4c9d 100644
--- a/test/irb/test_ruby_lex.rb
+++ b/test/irb/test_ruby_lex.rb
@@ -602,5 +602,10 @@ def foo
         pos_to_index[t.pos] = i
       }
     end
+
+    def teardown
+      h = { create: $count_create, to_a: $count_to_a, access: $count_access};
+      puts h.inspect
+    end
```

These were the results:

```
.{:create=>45,452, :to_a=>0, :access=>289,587}
```

The object is accessed 6.37x more often than it is created (289,587/ 45,452 # ~> 6.37x) . From my microbenchmark object access is 5.8x faster than creation (19.642M / 3.361M # ~> 5.8x). I don't know how close the irb tests are to real-world use.