@mame 2021-11-21
Thank you very much, what great this feedback is!

> it looks like master is slower by ~1.8%

I guess we'll have to look into this one. @ko1 has already added :eyes: emoji, so I believe he will take a look :trollface: