@schneems 2021-11-21
Sounds good. @ko1 let me know if I can help or give extra information. Also to be clear, I agree that the struct patch worked and makes master access faster, I don’t know why the delta exists between preview1.

Also if struct access can be sped up, maybe struct creation can be sped up as well and we wouldn’t need my patch. I don’t have the familiarity to understand optimizing that C code though. I don’t know if there’s a fundamental blocker or if it’s just never been attempted.