@mame 2021-11-24
Wait, [the test script](https://gist.github.com/schneems/1edfff044a082ca61f962066541dfdb4#file-gistfile1-txt) uses `array.rb`, but *the file was changed* after 3.1.0-preview1 https://github.com/ruby/ruby/commit/a24eeee5567a14841b37d9a3428e14e4f3c45c07 🤣 Did you make sure to use the same content file?