@mame 2021-11-24
> ## This patch
> Both class creation and `to_a` are still faster using a class instead of a struct:

According to @ko1's investigation, in terms of instance creation, the main difference between the current impl. has an overhead for calling `super` in `State#initialize`.

```
def initialize(i) super(i, Ripper.lex_state_name(i)).freeze end
```

After your patch, `initialize` simply assigns the arguments to instance variables, with no call to `super`. I didn't investigate `to_a`, but maybe there's something similar.

Anyway, I tried the current master (df8f8581153b4df7e7f042afd088b905f13b654c) with and without your patch, and the performance improvement is still observable: 17.7 sec. → 17.4 sec. It is never big, but maybe good. I'd like to merge it unless there is objection.