@schneems 2021-11-29
> Wait, the test script uses array.rb, but the file was changed after 3.1.0-preview1 a24eeee 🤣 Did you make sure to use the same content file? 

Oops, I did not 🙈 good catch.

>  I'd like to merge it unless there is objection.

It is okay with me. Please let me know if there are more questions or any requests for changes.