@yahonda 2021-12-06
@mame As discussed offline, this change cause a lot of warnings at Rails console. Please consider to address it. 

https://gist.github.com/yahonda/8cbb81e193b10f7f1be7f6ffb2c738e5